Current still broken, voltage works!

+++ Electrical/batController/firmware/libs/ina228/ina228.h
@@ -43,7 +43,7 @@ int ina228_reset(ina228_config *config);
 float ina228_voltage(ina228_config *config);
 uint32_t ina228_voltage_raw(ina228_config *config);
 float ina228_current(ina228_config *config);
-uint16_t ina228_current_raw(ina228_config *config);
+uint32_t ina228_current_raw(ina228_config *config);
 int reg_write(i2c_inst_t *i2c, 
                 const uint addr, 
                 const uint8_t reg,
